Update lcdc's phy_addr to compatible with 32 bit system (cherry picked from commit 7ff61a66bcb1e9cb79683e6b080ac4cb2e0a1fd6)

+++ include/hardware/rk_fh.h
@@ -36,7 +36,7 @@ struct rk_lcdc_post_cfg{
 struct rk_fb_area_par {
 	u8  data_format;        /*layer data fmt*/
 	short ion_fd;
-	unsigned long phy_addr;
+	unsigned int phy_addr;
 	short acq_fence_fd;
 	u16  x_offset;
 	u16  y_offset;
